Work around missing ads
IMA may send an error event if the bumper before / in an ad pod cannot be loaded. In that case, the ad pod position could be unknown and would be based on the "expected" position. That would lead to marking the whole ad break as an error and when the load event for the mid roll ads come in, the ads would be ignored.

This change delays the error if no ad pod index is known. If the index is known, the ad is marked as error immediately.